backend/requirements.txt
pandas==2.1.4
watchdog==3.0.0
fastapi==0.111.1
fastapi-cli==0.0.4
boto3==1.34.11
botocore==1.34.11
pydantic==2.5.2
pydantic_core==2.14.5
uvicorn[standard]==0.30.1
pyyaml==6.0
python-dotenv==1.0.0
requests==2.32.3
